Simplified theming and img refs
Implemented a new and simpler method to apply a custom color theme.
Theme now defined in one file and applied once to the <main>
DOM element. This change has far reaching consequences for app
structure and will lead to the simplification and even removal of a
number of atomic components.